+++ src/main/java/org/hankki/hankkiserver/api/auth/service/AuthService.java
@@ -16,11 +16,11 @@ import org.hankki.hankkiserver.domain.user.model.UserInfo;
 import org.hankki.hankkiserver.domain.user.model.UserStatus;
 import org.hankki.hankkiserver.event.EventPublisher;
 import org.hankki.hankkiserver.event.user.CreateUserEvent;
-import org.hankki.hankkiserver.external.openfeign.apple.AppleClientSecretGenerator;
-import org.hankki.hankkiserver.external.openfeign.apple.AppleOAuthProvider;
-import org.hankki.hankkiserver.external.openfeign.dto.SocialInfoDto;
-import org.hankki.hankkiserver.external.openfeign.kakao.KakaoOAuthProvider;
-import org.hankki.hankkiserver.external.openfeign.kakao.dto.KakaoUnlinkRequest;
+import org.hankki.hankkiserver.external.openfeign.login.apple.AppleClientSecretGenerator;
+import org.hankki.hankkiserver.external.openfeign.login.apple.AppleOAuthProvider;
+import org.hankki.hankkiserver.external.openfeign.login.dto.SocialInfoDto;
+import org.hankki.hankkiserver.external.openfeign.login.kakao.KakaoOAuthProvider;
+import org.hankki.hankkiserver.external.openfeign.login.kakao.dto.KakaoUnlinkRequest;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
