+++ src/main/java/org/hankki/hankkiserver/external/openfeign/login/apple/ApplePrivateKeyGenerator.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.external.openfeign.apple;
+package org.hankki.hankkiserver.external.openfeign.login.apple;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
