+++ src/main/java/org/hankki/hankkiserver/external/openfeign/login/kakao/dto/KakaoAccount.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.external.openfeign.kakao.dto;
+package org.hankki.hankkiserver.external.openfeign.login.kakao.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
