+++ src/main/java/org/hankki/hankkiserver/external/openfeign/login/kakao/dto/KakaoUserProfile.java
@@ -1,4 +1,4 @@
-package org.hankki.hankkiserver.external.openfeign.kakao.dto;
+package org.hankki.hankkiserver.external.openfeign.login.kakao.dto;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
